db: add tests for users (#6116)
* Add new methods

* Use Users.Create to replace previous hack

* Reduce side effect

* Do not clear tables when test failed

* test_users_Authenticate

* Rename constant

* test_users_Create

* test_users_GetByEmail

* test_users_GetByID

* test_users_GetByUsername

+++ internal/db/main_test.go
@@ -42,7 +42,11 @@ func TestMain(m *testing.M) {
 }
 
 // clearTables removes all rows from given tables.
-func clearTables(db *gorm.DB, tables ...interface{}) error {
+func clearTables(t *testing.T, db *gorm.DB, tables ...interface{}) error {
+	if t.Failed() {
+		return nil
+	}
+
 	for _, t := range tables {
 		err := db.Delete(t).Error
 		if err != nil {
